requirements.txt
Flask==3.0.3
redis==5.0.8  # python client for redis
mysql-connector-python==9.0.0  # MySQL driver written in Python
